requirements.txt
# =============================================================================
# Tieplm AI Assistant - Consolidated Requirements
# =============================================================================

# YouTube download
yt-dlp

# Audio/Video processing
ffmpeg-python

# Transcription (Open Source Whisper)
openai-whisper
torch
torchaudio

# FastAPI and server
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy
psycopg2-binary
alembic

# Vector DB
qdrant-client

# OpenAI (for embeddings and LLM)
openai

# RAG and Reranking
sentence-transformers  # For cross-encoder reranking
rank-bm25  # For BM25 lexical search

# Evaluation metrics
rouge-score
nltk

# Environment and config
python-dotenv
pydantic
pydantic-settings
pyyaml

# Utilities
httpx
tqdm
pandas

